requirements.txt
# ============================================
# BACKEND FRAMEWORK
# ============================================
fastapi==0.109.0
uvicorn[standard]==0.27.0

# ============================================
# DATABASE
# ============================================
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
greenlet>=3.0.0

# ============================================
# AUTHENTICATION
# ============================================
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0

# ============================================
# DATA VALIDATION
# ============================================
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# ============================================
# AI / ML
# ============================================
openai==1.12.0
chromadb==0.4.22
sentence-transformers==2.3.1
tiktoken==0.5.2

# ============================================
# DOCUMENT PROCESSING
# ============================================
pypdf==4.0.1

# ============================================
# CACHE
# ============================================
redis==5.0.1

# ============================================
# TESTING
# ============================================
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# ============================================
# UTILITIES
# ============================================
python-dotenv==1.0.0
loguru==0.7.2